requirements.txt
# Core Dependencies
streamlit>=1.31.1
openai>=1.109.1,<2.0.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# API and HTTP
requests>=2.31.0
urllib3>=2.1.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.1

# Date/Time
python-dateutil>=2.8.2

# Logging
colorlog>=6.8.2
